# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Port de Russell",
    "description": "API pour gérer les catways, les réservations et les utilisateurs du port de plaisance de Russell.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Serveur local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "username": { "type": "string", "example": "admin" },
          "email": { "type": "string", "example": "admin@example.com" },
          "password": { "type": "string", "example": "MotDePasse123!" }
        },
        "required": ["username", "email", "password"]
      },
      "Catway": {
        "type": "object",
        "properties": {
          "catwayNumber": { "type": "integer", "example": 1 },
          "catwayType": { "type": "string", "enum": ["long", "short"], "example": "long" },
          "catwayState": { "type": "string", "example": "En bon état" }
        },
        "required": ["catwayNumber", "catwayType", "catwayState"]
      },
      "Reservation": {
        "type": "object",
        "properties": {
          "catwayNumber": { "type": "integer", "example": 1 },
          "clientName": { "type": "string", "example": "John Doe" },
          "boatName": { "type": "string", "example": "Sea Breeze" },
          "startDate": { "type": "string", "format": "date", "example": "2025-11-01" },
          "endDate": { "type": "string", "format": "date", "example": "2025-11-05" }
        },
        "required": ["catwayNumber", "clientName", "boatName", "startDate", "endDate"]
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "example": "Description de l'erreur" }
        }
      }
    }
  },
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Connexion",
        "description": "Permet à un utilisateur de se connecter et de recevoir un token JWT.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "admin@example.com" },
                  "password": { "type": "string", "example": "MotDePasse123!" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Connexion réussie",
            "content": {
              "application/json": {
                "example": {
                  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                  "user": { "username": "admin", "email": "admin@example.com" }
                }
              }
            }
          },
          "401": {
            "description": "Email ou mot de passe invalide",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "get": {
        "summary": "Déconnexion",
        "description": "Déconnecte l'utilisateur et invalide le token JWT.",
        "responses": { "200": { "description": "Déconnexion réussie." } }
      }
    },
    "/users": {
      "get": {
        "summary": "Liste des utilisateurs",
        "description": "Retourne la liste de tous les utilisateurs.",
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/User" } }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Créer un utilisateur",
        "description": "Crée un nouvel utilisateur dans la base.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/User" }
            }
          }
        },
        "responses": {
          "201": { "description": "Utilisateur créé avec succès" },
          "400": {
            "description": "Erreur de validation",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
    "/users/{email}": {
      "get": {
        "summary": "Détails d'un utilisateur",
        "parameters": [
          { "name": "email", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Email de l'utilisateur à récupérer" }
        ],
        "responses": {
          "200": { "description": "Détails de l'utilisateur", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } },
          "404": { "description": "Utilisateur non trouvé", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "put": {
        "summary": "Modifier un utilisateur",
        "parameters": [
          { "name": "email", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/User" } }
          }
        },
        "responses": {
          "200": { "description": "Utilisateur modifié avec succès" },
          "400": { "description": "Erreur de validation", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "delete": {
        "summary": "Supprimer un utilisateur",
        "parameters": [
          { "name": "email", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Utilisateur supprimé avec succès" },
          "404": { "description": "Utilisateur non trouvé", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/catways": {
      "get": {
        "summary": "Liste des catways",
        "description": "Retourne la liste de tous les catways.",
        "responses": {
          "200": {
            "description": "Succès",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Catway" } } } }
          }
        }
      },
      "post": {
        "summary": "Créer un catway",
        "description": "Ajoute un nouveau catway dans la base.",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Catway" } } }
        },
        "responses": {
          "201": { "description": "Catway créé avec succès" },
          "400": { "description": "Erreur de validation", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/catways/{id}": {
      "get": {
        "summary": "Détail d'un catway",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "description": "Numéro du catway" }
        ],
        "responses": {
          "200": { "description": "Détails du catway", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Catway" } } } },
          "404": { "description": "Catway non trouvé", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "put": {
        "summary": "Modifier l'état d'un catway",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Catway" } } } },
        "responses": { "200": { "description": "Catway modifié avec succès" } }
      },
      "delete": {
        "summary": "Supprimer un catway",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Catway supprimé avec succès" },
          "404": { "description": "Catway non trouvé", "content": { "application/json/schema": { "$ref": "#/components/schemas/Error" } } }
        }
      }
    },
    "/catways/{id}/reservations": {
      "get": {
        "summary": "Liste des réservations d'un catway",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Renvoie les réservations", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Reservation" } } } } } }
      },
      "post": {
        "summary": "Créer une réservation",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Reservation" } } } },
        "responses": {
          "201": { "description": "Réservation créée avec succès" },
          "400": { "description": "Erreur de validation", "content": { "application/json/schema": { "$ref": "#/components/schemas/Error" } } }
        }
      }
    },
    "/catways/{id}/reservations/{idReservation}": {
      "get": {
        "summary": "Détails d'une réservation",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } },
          { "name": "idReservation", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Détails de la réservation", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Reservation" } } } } }
      },
      "put": {
        "summary": "Modifier une réservation",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } },
          { "name": "idReservation", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Reservation" } } } },
        "responses": { "200": { "description": "Réservation modifiée avec succès" } }
      },
      "delete": {
        "summary": "Supprimer une réservation",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } },
          { "name": "idReservation", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Réservation supprimée avec succès" },
          "404": { "description": "Réservation non trouvée", "content": { "application/json/schema" : { "$ref": "#/components/schemas/Error" } } }
        }
      }
    }
  }
}