# routes/reservations.global.routes.js
const express = require('express');
const Reservation = require('../models/Reservation');

const router = express.Router();

// GET all reservations
router.get('/', async (req, res) => {
  try {
    const reservations = await Reservation.find();
    res.json(reservations);
  } catch (error) {
    console.error(error);
    res.status(500).json({ message: 'Erreur serveur' });
  }
});

module.exports = router;